requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
keras_facenet==0.3.2
numpy~=1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
scikit_learn==1.4.2

opencv-python~=4.9.0.80
scikit-learn~=1.4.2
dash~=2.17.0
requests~=2.31.0